% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FindFiniteSet_L2}
\alias{FindFiniteSet_L2}
\title{Determine the finite set {k: P(L_i^{(2)} = k | C_i) >= v_i} for each spot i}
\usage{
FindFiniteSet_L2(v, c, K2_max, G, eta_k, a_eta, b_eta, psi2, geoq)
}
\arguments{
\item{v}{Augmented variable for spot i.}

\item{c}{Consensus clustering indicator for spot i.}

\item{K2_max}{Maximum in the gene-expression-specific cluster indicators L2.}

\item{G}{Number of genes (feature dimension of the gene expression matrix).}

\item{eta_k}{Mean value matrix for gene expression data. Rows: gene-expression cluster number. Columns: genes.}

\item{a_eta}{Mean of the normal prior for \eqn{\eta}.}

\item{b_eta}{Standard deviation of the normal prior for \eqn{\eta}.}

\item{psi2}{Gene-expression contribution weight.}

\item{geoq}{Hyperparameter of the geometric distribution ranging in (0,1).}
}
\value{
FindFiniteSet_L2 returns an R list including the following information.
\item{finiteSet}{vector, the finite set this function aims to find.}
\item{eta_k}{matrix, updated mean value matrix (maybe add some samples drawn from prior).}
\item{K2_max}{integer, updated value of K2_max (after possibly adding some samples).}
}
\description{
Determine the finite set {k: P(L_i^{(2)} = k | C_i) >= v_i} for each spot i
}
