% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SticksToPi}
\alias{SticksToPi}
\title{Convert beta random variables to weights pi_k in stick-breaking process}
\usage{
SticksToPi(sticks)
}
\arguments{
\item{sticks}{A beta random variable vector}
}
\value{
A vector of weights pi_k
}
\description{
Convert beta random variables to weights pi_k in stick-breaking process
}
