% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc_rbeta}
\alias{trunc_rbeta}
\title{Draw samples from truncated beta distribution}
\usage{
trunc_rbeta(num, s1, s2, w1, w2, by = 10^(-7))
}
\arguments{
\item{num}{Number of samples}

\item{s1}{Shape1}

\item{s2}{Shape2}

\item{w1}{Lower bound}

\item{w2}{Upper bound}

\item{by}{Distance between two nearby points}
}
\value{
A vector of samples
}
\description{
Draw samples from truncated beta distribution
}
